# spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wallet Tracker REST API",
    "version": "1.0.0",
    "description": "Contract for the explicit REST operations implemented by the current server: health, authentication, categories, transactions, synthetic batch import, schedules, summary, settings and AI drafts. Descriptions record implementation rules that JSON Schema cannot fully express. Document validation alone is not evidence of executed HTTP scenarios: check:openapi validates this contract and test:contract separately exercises disposable offline/demo servers, reporting observed and unexercised statuses. No full-project regression or GitHub CI result is implied by this document. OpenAPI 3.0.3 is selected for parser compatibility, not claimed to be the latest version. This describes a personal project, not a production service."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local application; PORT is configurable. No hosted demo is promised."
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/api/imports/preview": {
      "post": {
        "operationId": "previewSyntheticImport",
        "tags": ["Synthetic import"],
        "summary": "Validate and review a synthetic expense batch without saving",
        "description": "Authenticated caller-supplied JSON only; no external provider request. Exactly one fixed source, 1–10 records, strict numeric integer cents, valid local dates and exact own-category names. Existing external IDs are account-scoped and compared by semantic fingerprint. Preview writes neither expenses nor retry-ledger rows; confirm is separately required. Provisional choices are recorded in qa/docs/analysis-synthetic-import.md.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyntheticImportRequest" } } } },
        "responses": {
          "200": { "description": "Read-only review of each submitted record.", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyntheticImportPreview" } } } },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/KeyedConflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/imports/confirm": {
      "post": {
        "operationId": "confirmSyntheticImport",
        "tags": ["Synthetic import"],
        "summary": "Atomically save a reviewed synthetic expense batch",
        "description": "Caller must explicitly confirm; the API does not require a preview token and revalidates the entire batch. One BEGIN IMMEDIATE transaction covers all records and existing keyed-save ledger entries. Same external ID and content replays the original transaction; changed content or deleted original returns 409 without a partial batch. This does not connect to an external provider.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyntheticImportRequest" } } } },
        "responses": {
          "200": { "description": "Entire batch replayed without new expenses.", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyntheticImportConfirmation" } } } },
          "201": { "description": "At least one new expense saved atomically.", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyntheticImportConfirmation" } } } },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/KeyedConflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT returned by registration/login. The application's Authorization header uses case-sensitive Bearer followed by a token. Missing, invalid or expired token returns 401. The three public operations explicitly override this root requirement with security=[]."
      }
    },
    "schemas": {
      "SyntheticImportRequest": {
        "type": "object", "additionalProperties": false, "required": ["source", "records"],
        "properties": {
          "source": { "type": "string", "enum": ["synthetic-demo-v1"] },
          "records": { "type": "array", "minItems": 1, "maxItems": 10, "items": { "$ref": "#/components/schemas/SyntheticImportRecord" } }
        }
      },
      "SyntheticImportRecord": {
        "type": "object", "additionalProperties": false,
        "required": ["external_id", "amount_cents", "spent_on", "category_name"],
        "properties": {
          "external_id": { "type": "string", "pattern": "^[A-Za-z0-9_-]{1,64}$" },
          "amount_cents": { "type": "integer", "minimum": 1, "maximum": 100000000 },
          "spent_on": { "$ref": "#/components/schemas/LocalDate" },
          "category_name": { "type": "string", "minLength": 1, "maxLength": 60 },
          "note": { "type": "string", "nullable": true, "maxLength": 160 }
        }
      },
      "SyntheticImportPreview": {
        "type": "object", "additionalProperties": false, "required": ["source", "records"],
        "properties": {
          "source": { "type": "string", "enum": ["synthetic-demo-v1"] },
          "records": { "type": "array", "minItems": 1, "maxItems": 10, "items": {
            "type": "object", "additionalProperties": false,
            "required": ["external_id", "amount_cents", "spent_on", "category_name", "note", "category_id", "state", "transaction_id"],
            "properties": {
              "external_id": { "type": "string" }, "amount_cents": { "type": "integer", "minimum": 1 },
              "spent_on": { "$ref": "#/components/schemas/LocalDate" }, "category_name": { "type": "string" },
              "note": { "type": "string", "nullable": true }, "category_id": { "type": "integer", "nullable": true, "minimum": 1 },
              "state": { "type": "string", "enum": ["new", "already_imported"] },
              "transaction_id": { "type": "integer", "nullable": true, "minimum": 1 }
            }
          } }
        }
      },
      "SyntheticImportConfirmation": {
        "type": "object", "additionalProperties": false, "required": ["source", "records"],
        "properties": {
          "source": { "type": "string", "enum": ["synthetic-demo-v1"] },
          "records": { "type": "array", "minItems": 1, "maxItems": 10, "items": {
            "type": "object", "additionalProperties": false,
            "required": ["external_id", "transaction_id", "replayed"],
            "properties": {
              "external_id": { "type": "string" }, "transaction_id": { "type": "integer", "minimum": 1 },
              "replayed": { "type": "boolean" }
            }
          } }
        }
      },
...
      "ErrorMessage": { "type": "string", "minLength": 1 },
      "Error400": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["VALIDATION_FAILED"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
      "Error401": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["UNAUTHORIZED"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
      "Error404": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["NOT_FOUND"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
...
      "KeyedError409": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["IDEMPOTENCY_CONFLICT", "IDEMPOTENCY_REPLAY_UNAVAILABLE"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
      "Error413": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["PAYLOAD_TOO_LARGE"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
...
      "Error500": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["INTERNAL"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
...
      "LocalDate": {
        "type": "string",
        "format": "date",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
        "description": "Real local calendar date YYYY-MM-DD, no timezone conversion. Source validates month/day including leap years and imposes no additional year ceiling within four digits."
      },
...
    },
    "parameters": {
...
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "A lowercase UUID matching the response header pattern is echoed. Any other value, including uppercase or non-UUID text, is replaced by a generated UUID; it does not cause 400."
      }
    },
    "headers": {
...
      "RequestId": {
        "description": "Always present. Caller lowercase UUID is echoed; otherwise a new UUID is generated. The x-required extension is for the later response validator, not an OpenAPI required keyword.",
        "x-required": true,
        "schema": {
          "type": "string",
          "format": "uuid",
          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
        }
      },
...
    },
    "responses": {
...
      "KeyedConflict": {
        "description": "IDEMPOTENCY_CONFLICT for changed canonical payload or IDEMPOTENCY_REPLAY_UNAVAILABLE when the original expense was deleted.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KeyedError409" } } }
      },
...
      "NotFound": {
        "description": "NOT_FOUND: no requested category, transaction or schedule belongs to the account. Missing and foreign owned IDs both return404, not403.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } }
      },
...
      "InvalidRequest": {
        "description": "VALIDATION_FAILED: invalid values, date range, ID conversion or PATCH shape, or malformed JSON rejected before a handler. Login bad/missing credential fields instead use401. The global parser also applies before auth and GET handlers.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } }
      },
      "Unauthorized": {
        "description": "UNAUTHORIZED: login uses the same E-mail or password is incorrect message for absent/non-string/rejected credentials. Protected mounts refuse absent/malformed bearer headers or invalid/expired tokens with their authentication message. No account-existence disclosure.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error401" } } }
      },
...
      "PayloadTooLarge": {
        "description": "PAYLOAD_TOO_LARGE: the global Express JSON parser's default 100 KiB body limit is exceeded.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error413" } } }
      },
...
      "InternalError": {
        "description": "INTERNAL: Unexpected server error. Documented possible unhandled failure; not induced or observed by this document-validation slice.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error500" } } }
      }
    }
  }
}